Handlers: Refactor condition checkers into own module
